torch_baseline/requirement.txt
transformer==2.11.0
tokenizers==0.7.0
torch==1.7.0


